First example with VN_Messages done.
First example with VN_Messages on how we might work with them.
The problem that is left to solve is if we want fixed sized messages for all
messages. This would mean that we could easier convert between message types but
at the cost of a lot of unchecked conversions.