Added template for VN Message Logger
The idea is that instead of typing debug message all over the place it easier
just to call a logging function. This logging function can easily be swapped
out between different versions such as an empty logger for no logging, a print
out version for only print outs to stdout or perhaps logging to file.

It may also be possible to create a composite of these to log information to
different outputs.

The main concept here is that your main priority is to log where different
messages show up in the system. To log other aspects, another type of logger
is needed.