This causes routing to fail.
When CAS sends the Assign_Address_Block the SM-L never receives it in the
application layer. Instead the CAS gets the message back and receives it.

The problem is probably because the Protocol_Routing is not started until the
SM-L sends its first message from the application layer. Perhaps it is possible
to "start" protocol routing before the SM-L sends its first message?